Create a function to connect to a data base in memory for tests
