+++ sonar-project.properties
@@ -16,4 +16,4 @@ sonar.javascript.lcov.reportPaths=./coverage/lcov.info
 
 sonar.test.inclusions=**/**/*.test.ts
 
-sonar.coverage.exclusions=src/database/index.ts,src/index.ts,src/server/startServer.ts
+sonar.coverage.exclusions=src/database/index.ts,src/index.ts,src/server/startServer.ts,src/testsSetup/connectDataBaseTests.ts
