+++ src/server/index.ts
@@ -5,7 +5,6 @@ import handleEndpointNotFound from "./errors/handleEndpointNotFound.js";
 import handleGeneralErrors from "./errors/handleGeneralErrors.js";
 
 const app = express();
-
 app.disable("x-powered-by");
 
 app.use(morgan("dev"));
@@ -13,7 +12,6 @@ app.use(morgan("dev"));
 app.use("/games", gamesRouter);
 
 app.use(handleEndpointNotFound);
-
 app.use(handleGeneralErrors);
 
 export default app;
